@@ -4,18 +4,18 @@
     "type":        "library",
     "license":     "MIT",
 
-
-
     "authors": [
         {
             "name":     "Sid Roberts",
             "email":    "sid@sidroberts.co.uk",
             "homepage": "https://sidroberts.co.uk",
             "role":     "Developer"
+        },
+        {
+            "name":     "Jujudje",
+            "role":     "Contributor"
         }
     ],
-
-
 
     "support": {
         "source": "https://github.com/SidRoberts/cron",
@@ -23,22 +23,16 @@
         "wiki":   "https://github.com/SidRoberts/cron/wiki"
     },
 
-
-
     "require": {
-        "php": "~7.1",
-
-        "dragonmantank/cron-expression": "~1.0"
+        "php": "^8.0",
+        "dragonmantank/cron-expression": "3.0.1"
     },
 
     "require-dev": {
-        "codeception/codeception": "~4.0",
-        "codeception/module-asserts": "~1.0",
-
-        "squizlabs/php_codesniffer": "~3.4"
+        "codeception/codeception": "5.0.7",
+        "codeception/module-asserts": "3.0.0",
+        "squizlabs/php_codesniffer": "3.7.1"
     },
-
-
 
     "autoload": {
         "psr-4": {
